removed equality_tolerance from global opts structure and replaced it with length units per meter so that stacking height eps can use that
Previously, the global options structure had an internal eps value that wasn't really actually tweakable. It has been replaced with a length_units_per_meter so that the stacking algorithm could use that factor to establish a surface matching epsilon. This was necessary because with models in millimeters the surface matching epsilon could get too small to be practical. The global epsilon has been replaced by a constant EPS_MAGIC, which more accurately describes it.