skipped element subtraction if the two elements share a plane opposite
When two elements "intersect" as a face, the volume subtraction takes forever but will eventually be a no-op. There is now a check to determine when this is happening and skip the subtraction. This means that extremely pathological cases will yield a false negative (if two elements opposite-share a face but actually intersect in their volumes elsewhere) but I'm not worried about these cases.