envelope calculation will be skipped if all other surfaces are too far away
When determining whether to perform an envelope calculation, all non-base parallel surfaces will be checked for how far away they are from the base. If they're all further than a cutoff parameter the envelope calculation will be skipped and the base set up as a halfblock. This saves time in blocking (4% on the BLIS building).