rearranged operation status priorities so that problems trump InProgress
Operations must report their status. Previously, if an operation was in progress it would report that even if it warnings or errors. The ordering has been changed so that in-progress operations report the presence of warnings or errors immediately. This allows the application to display the appropriate icon in the taskbary overlay without waiting for the operation to complete so that the user knows of the problem as soon as possible.