removed unused "material usage" code
There used to be machinery to track construction usage so that mappings wouldn't be required for materials that didn't end up in any space boundary layer sets. This hasn't been displayed to the user for awhile, and would otherwise have needed to be updated when support for material lists was added, so I just deleted it.