complex output construction retooling
The code for generating IDF constructions has been reworked to be more readable, give (slightly) better warnings, and remove an old code path that would just arbitrarily make a choice and then not warn the user about it. Specifically, previously, composites in the IFC were mapped directly to the IDF, thicknesses and everything. Now, the generation will not map it at all (and emit a warning), making the user choose themselves what to do.