multiview_solids from extruded_area_solids now check to make sure that their extrusion direction isn't parallel to their base
This is only currently happening because the Ifc Adapter rewrite isn't correct yet, but it's a good check to add just in case (the alternative is the poly generation just failing and taking the whole run down with it). If an element has this property is now ignored and a warning emitted.