solids in the ifc adapter now track axis1 and axis2 (and therefore axis3) if they have them, instead of just axis1.
This is another change to enable addition of composite layer direction tracking in the future. Now solids have a sense of all their axes, not just axis1. Previously, if the composite axis had been axis2 (for example) there would have been no way to find out what that axis is.