openings now bypass the stacking step
Openings now do not get stacked. Instead, they're assigned directly to space boundaries that come out of stacking. This is for two reasons. First, the stacking code will need a major refactor soon (to conform to the paper's description for performance reasons) and so simplifying it is necessary. Second, stacking openings is almost always unnecessary and takes time. The downside is that openings that span multiple spaces won't get handled correctly now. I think the tradeoff is worth it.