added code to account for breps with adjacent faces
Revit has started doing this weird "triangulating" thing where it splits up faces arbitrarily (I think this was a byproduct of them fixing their normal issue). The polyhedra code needed to be modified to accommodate that. It's causing a test to break, although the test was "this incorrect input fails" so I haven't decided what to do about it.