prevented nef polygon snapping when the target nef has more than eight vertices
Snapping is an asymptotically problematic process right now, and I don't have to figure out a better approach at the moment so I'm just capping the complexity at which it occurs (the original problem was simple rectangles in other rectangles, not these behemoth floor slabs that are causing me to notice this). The choice of eight vertices is arbitrary. I hope this doesn't ruin anything.