conversion to interface structures now takes a variable epsilon as an argument, and doesn't assume that final geometry cleaning succeeds
The variable epsilon is to account for units scaling, and the geometry cleaning is checked so that tiny space boundaries get pruned. This is both because tiny space boundaries just extend the energy simulation for not much gain and in certain cases, when opposite space boundaries are cleaned identically, bad IDFs are generated.