added a test to make sure that a newly-created equality context will snap near-axis directions to the axes
Previously, this wasn't actually guaranteed, and it always just happened to work because the axes were the first directions requested. I've run into a model where this wasn't the case though. It looks like the actual commit to fix the problem slipped through earlier with the wrong commit message (because I am a Bad Developer) but here's a unit test to capture the behavior.