added simple_face::is_planar
If a simple face is constructed from non-coplanar points (or specifically, points that can't be snapped to an axis plane), invariants down the line are broken. (This specific case was discovered during poly assembly, where non-planar faces yield bad nef polyhedra). I've added a check for this and a test. I haven't actually fixed the problem behavior yet.