modified simple_face constructor to enforce coplanarity
Previously, simple faces never guaranteed that their points were all coplanar. This was usually not a problem because they either were coplanar incidentally or were going straight into an oriented_area, which do enfore planarity. However, I've been running into some models where this was not the case and this was causing problems.