changed all the "geomery too complicated" language into "stack overflow" language
"Geometry too complicated" was an artifact of our previous numeric library that liked to do lots of work on the stack and would therefore overflow semi-regularly. LEDA doesn't do this, but stack overflows will still bring the app down if they aren't handled, so I still catch them. This way the message is just a little less confusing for the user.