removed assertion
The gist of this change is explained in the committed comment. Basically, if the columns of a transformation matrix are not orthonormal it won't transform polygons in 3D the same way it transforms direction_3s (I'm not quite sure why, but this is what the CGAL mailing list told me). This situation can arise because IfcBuildAxes isn't exact. However, IfcBuildAxes produces axes that are always close to orthonormal, so this isn't actually a problem in practice (because the direction snapping in the core should be handling this downstream anyway).