added a force_planar parameter to the normal simple_face constructor
In some cases, simple_faces should be forced to be planar when constructed (specifically when they're used as the base of extrusions). In other cases, they shouldn't (when they're brep faces). In this latter case, converting the brep to a nef solid won't work, but that's fine as long as we don't do that (and go directly to oriented_areas instead). I've modified the constructor to be able to handle both of these scenarios.