changed internal representation of multiview_solids away from a variant type
Previously, multiview_solids were internally represented with a variant type that was updated whenever a new "form" was needed. The problem with this was that some conversions were one-way, and data about the solid's geometry could be lost. The new method caches and keeps each representation independently so none is lost.