replaced public poly_builder constructors with static factory functions
I'm planning to modify poly_builders to be able to fail. I'd prefer not to add another exception type, so instead, I'm using the "optional return to indicate failure" method which is better in basically every way. However, the failure conditions themselves are not yet implemented, so the factory functions don't return boost::optionals just yet.