modified bg_vertex_data to store vertex area explicitly rather than retrieving it from the underlying data_ member on-demand
This came about because I was experimenting with flattening the areas. I didn't end up doing that because it didn't seem to impact performance at all (and unnecessary flattening is a bad idea), but I left the framework in in case I need to modify this stuff later. It has the side benefit of vastly simplifying the area getter (a()).