removed wrapped_nef_polygon snapping
wrapped_nef_polygon snapping has always been terrible (incredibly inefficient and of dubious use). I can make all the unit tests pass with a call to util::clean, and while that might be a condemnation of the unit tests, I feel perfectly justified making the change. If it turns out that snapping needs to be revisited I'll do it correctly (efficiently).