added assertion to check the output from the ifc adapter (w/r/t element ids)
The problem, that I've since discovered, is that the "next element id" is auto-incremented even when elements are dropped, causing the element ids to fall out of sync with their respective elements' positions in the elements array. This problem is not yet fixed, but now it's more noticeable when it happens.