big restructure of the project organization
Basically I've created a new top-level directory for everything, with all the source code now in src/. Additionally, all the project's dependencies now live either in dependencies/ (CGAL, Idf Toolbox, gmp, gtest) or are pointed to by new user macros. These user macros will need to be setup in a new property sheet if others wish to build the code (property sheets are not supposed to be version-controlled for this reason). Finally, all the dll-copying that needs to happen is handled by a variety of post-build events. The upshot of this is that anyone can clone, point to their personal copies of boost, LEDA, and EDM, build, and go. (I'm still not sure what to do about the fact that nobody will *have* a copy of EDM without having dropped ten grand previously.)