added null check for sbtbuilding geometry report
The Debug tab shows geometric information about the current SBT building, but if there is no current SBT building, these used to throw because of the null. (This didn't really matter because the default MVVM stuff just swallows the exceptions I guess.) It was annoying when debugging, so I've put a check in to return 0 if there's no building.