removed some legacy facade stuff from the ifc adapter
Originally, the exact geometry structures were simply versions of the interface structures with exact numbers instead of numbers. Then, I added real-class versions of faces, breps, etc. because I needed real functionality, but I didn't completely eliminate the old versions because they were still used in the Nef stuff which was awful and I didn't want to touch it. I'm finding myself needing to prune breps with voids (because they crash the core apparently), and making that change with two parallel versions of the geometry code seemed like a bad idea. So, I've eliminated the old versions. The Nef code is still around because I'm not quite up for refactoring that yet.