added some initonly qualifiers
I forget about these, but I might as well add them in. I tried with the native members (the pointers) as well but that interacts really poorly with const-correctness and dereferencing or something and since the whole point of initonly is to save me time I'm not going to try to track the problem down.