removed a const qualifier to allow variant dispatch to work
Previously, ifc_object::as_app_instance was const-qualified. This means that when this was passed to the static visitor, it wouldn't get picked up by the (non-const) Application_instance handler, even when it should have been. I've removed the const qualification so it works now. (The whole point of Application_instances is that they aren't const).