dependency/edm wrapper fiddling
The solution setup that allows people to use the Edm Wrapper is somewhat complicated and difficult to test properly. I had an opportunity to perform such a test, and some changes needed to be made. It's essentially fixing up build events and some .vcxproj trickery with the Edm Wrapper project.