changed viewmodel internal date handling
The previous way of doing it would be to try to construct dates and simply swallow a bad date exception (initially, the months and days are 0, but it doesn't matter). This made debugging super annoying (the debugger would capture every exception), so I changed it to a pass through a String.Format and a DateTime.TryParse.