added check for models lacking elements or spaces
This check happens *after* element extraction, which can be an expensive process. This kind of sucks, but the only way around it I can see is adding complexity to the interface. The problem is that a user could legitimately want the elements (or spaces) extracted even if spaces (or elements) aren't present, and there's no current way to signal this to the IFC adapter. So right now, it will just grab them both every time. If someone complains I can look into changing it.