fixed traversal to cause failing test to pass
The way transmission identification works is by subtracting areas from space faces so they don't get repeated. However, they were getting repeated. The root cause was that a structure wrapping the space faces was caching their areas, but the update was bypassing this cache, so it didn't ever have any effect. According to a comment, the only purpose for the cache was for potential future changes, so I just got rid of it.