use new custom properties for dep configuration
To manage dependencies we don't redistribute, I used to use user
property sheets. The problem with doing this was that fresh clones
didn't have the sheets, so none of the projects that required them
would load. I've discovered a way of specifying custom user properties
for things like this, so that has now been implemented. The README has
been updated to reflect this.