properly synchronize element ids with indices
The element_info "id" field actually indicates a construction id, and
needs to match the element_info's position in the elements array in
order for material layer lookups to succeed. The old system,
incomprehensibly, relied on tracking the "next" index manually, and
broke in the presence of shading elements. (I'm not sure how this was
never noticed before.) The id is now explicitly set to match the index.